clean up sac implementation
