--- offpolicy/algorithms/algorithm.py
@@ -1,68 +1,0 @@
-import abc
-
-
-class Algorithm(abc.ABC):
-
-    @abc.abstractmethod
-    def train(self, i, obs, act, reward, done, next_obs):
-        """Perform a single gradient descent update on the agent
-        using a batch of data sampled from a replay buffer
-
-        Args:
-
-        i: tf.dtypes.int64
-            the scalar training iteration the agent is currently on
-        obs: tf.dtypes.float32
-            a tensor shaped [batch_size, obs_size] containing observations
-        act: tf.dtypes.float32
-            a tensor shaped [batch_size, act_size] containing actions
-        reward: tf.dtypes.float32
-            a tensor shaped [batch_size, 1] containing a reward
-        done: tf.dtypes.bool
-            a tensor shaped [batch_size, 1] containing a done signal
-        next_obs: tf.dtypes.float32
-            a tensor shaped [batch_size, obs_size] containing observations
-        """
-
-        return NotImplemented
-
-    @abc.abstractmethod
-    def get_diagnostics(self, i, obs, act, reward, done, next_obs):
-        """Gather diagnostic information from the learning algorithm,
-        and return a dict containing tensors
-
-        Args:
-
-        i: tf.dtypes.int64
-            the scalar training iteration the agent is currently on
-        obs: tf.dtypes.float32
-            a tensor shaped [batch_size, obs_size] containing observations
-        act: tf.dtypes.float32
-            a tensor shaped [batch_size, act_size] containing actions
-        reward: tf.dtypes.float32
-            a tensor shaped [batch_size, 1] containing a reward
-        done: tf.dtypes.bool
-            a tensor shaped [batch_size, 1] containing a done signal
-        next_obs: tf.dtypes.float32
-            a tensor shaped [batch_size, obs_size] containing observations
-
-        Returns:
-
-        diagnostics: dict
-            a dict containing tensors whose statistics will be summarized
-        """
-
-        return NotImplemented
-
-    @abc.abstractmethod
-    def get_saveables(self):
-        """Collects and returns stateful objects that are serializeable
-        using the tensorflow checkpoint format
-
-        Returns:
-
-        saveables: dict
-            a dict containing stateful objects compatible with checkpoints
-        """
-
-        return NotImplemented
