--- offpolicy/networks/__init__.py
@@ -1,4 +1,0 @@
-from offpolicy.networks.feed_forward import FeedForward
-from offpolicy.networks.q_function import QFunction
-from offpolicy.networks.tanh_gaussian import TanhGaussian
-from offpolicy.networks.gaussian import Gaussian
